Final Project- Mad Libs Generator
During long trips, my mom would do word searches, while my brother and I would fill our time playing Mad Libs. A fun, quirky game where you fill in the blanks to finish off a story. These blanks can be anything from a number to a holiday, a plural known or an adjective. For my final project, I chose to create a game of Mad Libs where the user can choose between 1 or 3 different prompts and the game will ask you for different words to finish off the story. 
For this, I focused on utilizing the print function, input function and if functions to allow me to focus on this project. These correlate to module 2, and 3 from the summer semester which focus on statements and functions and allow the stored words to be utilized to finish off the story at the end.